CouponAPI/requirements.txt
asgiref==3.3.4
astroid==2.5.6
Babel==2.9.1
boto3==1.17.91
botocore==1.20.91
colorama==0.4.4
Django==3.0.5
django-crispy-forms==1.11.2
django-environ==0.4.5
django-money==2.0
django-moneyfield==0.2.1
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-phonenumber-field==5.1.0
django-rest-framework==0.1.0
django-smtp-ssl==1.0
django-storages==1.11.1
djangorestframework==3.12.4
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
isort==5.8.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mongoengine==0.23.1
phonenumbers==8.12.24
Pillow==8.2.0
psycopg2-binary==2.8.6
py-moneyed==1.2
pylint==2.8.2
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
urllib3==1.26.5
wrapt==1.12.1
